// src/constants/enums/turrets.ts
export enum TurretType {
    Chaingun = "Chaingun",
    Bolter = "Bolter",
    Laser = "Laser",
    PlasmaGun = "PlasmaGun",
    Cannon = "Cannon",
    RocketLauncher = "RocketLauncher",
    Railgun = "Railgun",
    Repair = "Repair",
    MiningLaser = "MiningLaser",
    RawMiningLaser = "RawMiningLaser",
    SalvagingLaser = "SalvagingLaser",
    RawSalvagingLaser = "RawSalvagingLaser",
    ForceGun = "ForceGun",
    TeslaGun = "TeslaGun",
    LightningGun = "LightningGun",
    PulseCannon = "PulseCannon",
    PointDefenseCannon = "PointDefenseCannon",
    PointDefenseLaser = "PointDefenseLaser",
    AntiFighter = "AntiFighter",
}