// src/constants/enums/components.ts
export enum ComponentType {
    Aluminum = "Aluminum",
    AmmunitionM = "AmmunitionM",
    AmmunitionS = "AmmunitionS",
    Conductor = "Conductor",
    Copper = "Copper",
    Crystal = "Crystal",
    ElectroMagnet = "ElectroMagnet",
    ElectromagneticCharge = "ElectromagneticCharge",
    EnergyCell = "EnergyCell",
    EnergyContainer = "EnergyContainer",
    EnergyInverter = "EnergyInverter",
    EnergyTube = "EnergyTube",
    ExplosiveCharge = "ExplosiveCharge",
    ForceGenerator = "ForceGenerator",
    Fuel = "Fuel",
    GaussRail = "GaussRail",
    Gold = "Gold",
    HighCapacityLens = "HighCapacityLens",
    HighPressureTube = "HighPressureTube",
    IndustrialTeslaCoil = "IndustrialTeslaCoil",
    LaserCompressor = "LaserCompressor",
    LaserHead = "LaserHead",
    LaserModulator = "LaserModulator",
    Lead = "Lead",
    MilitaryTeslaCoil = "MilitaryTeslaCoil",
    Nanobot = "Nanobot",
    PlasmaCell = "PlasmaCell",
    PowerUnit = "PowerUnit",
    Rocket = "Rocket",
    Servo = "Servo",
    Steel = "Steel",
    SteelTube = "SteelTube",
    TargetingCard = "TargetingCard",
    Transformator = "Transformator",
    Warhead = "Warhead",
    Wire = "Wire",
    Zinc = "Zinc",
}